target/requirements.txt
Flask
Flask-WTF
Jinja2
Flask-DebugToolbar
flask-cors
pyyaml==3.13
